Moved selectSAs function to utils, updated header date
